requirements.txt
Pygments==1.6rc1
argparse==1.2.1
-e git+git@github.com:bitly/bitly-api-python.git@11a2d9cfeddfc3361e31dec8d64e57b3280bfbda#egg=bitly_api-dev
distribute==0.6.24
fancycompleter==0.4
ipython==0.13.1
numpy==1.6.2
matplotlib==1.2.0
nose==1.2.1
pdbpp==0.7.2
pymongo==2.4.2
pyrepl==0.8.4
python-dateutil==2.1
six==1.2.0
wmctrl==0.1
wsgiref==0.1.2
